+++ lib/src/physics/clamping_with_overscroll_physics.dart
@@ -40,6 +40,9 @@ class ClampingWithOverscrollPhysics extends ClampingScrollPhysics {
 
   /// Creates scroll physics that prevent the scroll offset from exceeding the
   /// bounds of the content while handling the overscroll.
+  ///
+  /// The [state] must be the same [IndicatorController] that is passed to
+  /// the indicator widget, otherwise the behavior will be unpredictable.
   const ClampingWithOverscrollPhysics({
     super.parent,
     required ClampingWithOverscrollPhysicsState state,